+++ src/app/search-form/journey-point-input/journey-point-input.component.ts
@@ -1,11 +1,12 @@
-import { debounceTime, distinctUntilChanged } from 'rxjs/operators'
-import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { FormControl, FormGroupDirective, NgForm, Validators } from '@angular/forms';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators'
 
-import { SbbAutocompleteSelectedEvent } from '@sbb-esta/angular/autocomplete';
+import { SbbAutocompleteSelectedEvent, SbbAutocompleteTrigger } from '@sbb-esta/angular/autocomplete';
 import { SbbErrorStateMatcher } from '@sbb-esta/angular/core';
 
 import * as OJP_Types from 'ojp-shared-types';
+import * as OJP_Next from 'ojp-sdk-next';
 
 import { OJP_VERSION } from '../../config/constants';
 
@@ -15,8 +16,29 @@ import { UserTripService } from '../../shared/services/user-trip.service';
 import { AnyPlace, PlaceBuilder } from '../../shared/models/place/place-builder';
 import { PlaceLocation } from '../../shared/models/place/location';
 import { OJPHelpers } from '../../helpers/ojp-helpers';
+import { AnyLocationInformationRequest } from '../../shared/types/_all';
+import { GeoPositionBBOX } from '../../shared/models/geo/geoposition-bbox';
 
-type MapLocations = Record<OJP_Types.PlaceTypeEnum, AnyPlace[]>;
+interface RenderPlaceResult {
+  place: AnyPlace,
+  type: 'around_me' | 'place',
+  caption: string,
+  distance?: number,
+};
+
+const AroundMePlaceResult: RenderPlaceResult = (() => {
+  const place = PlaceLocation.Empty('Current Location');
+
+  const placeResult: RenderPlaceResult = {
+    place: place,
+    type: 'around_me',
+    caption: place.placeName,
+  };
+
+  return placeResult;
+})();
+
+type MapPlaces = Record<OJP_Types.PlaceTypeEnum, RenderPlaceResult[]>;
 type OptionLocationType = [OJP_Types.PlaceTypeEnum, string];
 
 export class ErrorStateMatcher implements SbbErrorStateMatcher {
@@ -31,21 +53,50 @@ export class ErrorStateMatcher implements SbbErrorStateMatcher {
   templateUrl: './journey-point-input.component.html',
   styleUrls: ['./journey-point-input.component.scss']
 })
-export class JourneyPointInputComponent implements OnInit, OnChanges {
-  private shouldFetchNewData = true;
+export class JourneyPointInputComponent implements OnInit {
+  @ViewChild(SbbAutocompleteTrigger, { static: true }) autocompleteInputTrigger: SbbAutocompleteTrigger | undefined;
 
-  public inputControl = new FormControl('', [Validators.required]);
+  private ignoreInputChanges: boolean;
 
-  public mapLookupPlaces: MapLocations;
+  public inputControl: FormControl;
+
+  public mapLookupPlaces: MapPlaces;
   public optionLocationTypes: OptionLocationType[];
 
   @Input() placeholder: string = '';
+  @Input() filterPlaceType: OJP_Types.PlaceTypeEnum | undefined;
   
-  @Input() currentPlace: AnyPlace | null;
+  private _currentRenderPlaceResult: RenderPlaceResult | null = null;
+  @Input() set place(place: AnyPlace | null) {
+    if (place) {
+      const placeName = place.computeName();
+      this.inputControl.setValue(placeName, { emitEvent: false });
+
+      this._currentRenderPlaceResult = {
+        place: place,
+        type: 'place',
+        caption: placeName,
+      };
+
+      this.mapLookupPlaces[place.type] = [this._currentRenderPlaceResult];
+    } else {
+      this._currentRenderPlaceResult = null;
+      this.mapLookupPlaces['stop'] = [AroundMePlaceResult];
+    }
+  }
+
   @Output() selectedNewPlace = new EventEmitter<AnyPlace>();
 
+  public useSingleSearchPool: boolean;
+
+  private sdk: OJP_Next.AnySDK; 
+
   constructor(private mapService: MapService, private userTripService: UserTripService, private languageService: LanguageService) {
-    this.mapLookupPlaces = {} as MapLocations;
+    this.ignoreInputChanges = false;
+    
+    this.inputControl = new FormControl('', [Validators.required]);
+
+    this.mapLookupPlaces = {} as MapPlaces;
     this.resetMapPlaces();
 
     this.optionLocationTypes = [
@@ -55,53 +106,63 @@ export class JourneyPointInputComponent implements OnInit, OnChanges {
       ['address', 'Addresses'],
     ];
 
-    this.currentPlace = null;
+    this.place = null;
+
+    this.useSingleSearchPool = false;
+
+    this.sdk = this.userTripService.createOJP_SDK_Instance(this.languageService.language);
   }
 
   ngOnInit() {
+    this.useSingleSearchPool = this.filterPlaceType !== undefined;
+
     this.inputControl.valueChanges.pipe(
       debounceTime(500),
       distinctUntilChanged()
-    ).subscribe((searchTerm: string | null) => {
-      if (searchTerm === null) {
-        return;
-      }
-
-      if (!this.shouldFetchNewData) {
-        return;
-      }
-
-      if (searchTerm.trim().length < 1) {
-        return;
-      }
-
-      const coordsPlace = PlaceLocation.initFromLiteralCoords(searchTerm);
-      if (coordsPlace) {
-        this.resetMapPlaces();
-        this.handleSelectedPlace(coordsPlace);
-        return;
-      }
-
-      this.fetchJourneyPoints(searchTerm);
+    ).subscribe((_) => {
+      this.onInputChangeAfterIdle();
     });
   }
-
-  ngOnChanges(changes: SimpleChanges) {
-    if ('currentPlace' in changes) {
-      const place = changes['currentPlace'].currentValue as AnyPlace;
-      if (place) {
-        this.inputControl.setValue(place.computeName(), { emitEvent: false });
-      }
+  
+  private onInputChangeAfterIdle() {
+    if (this._currentRenderPlaceResult?.type === 'around_me') {
+      return;
     }
+
+    const searchTerm = this.inputControl.value.trim();
+
+    if (this.ignoreInputChanges) {
+      return;
+    }
+
+    if (searchTerm.trim().length < 1) {
+      this.autocompleteInputTrigger?.openPanel();
+
+      this.useSingleSearchPool = true;
+      
+      this.resetMapPlaces();
+      this.mapLookupPlaces['stop'] = [AroundMePlaceResult];
+
+      return;
+    }
+
+    this.useSingleSearchPool = this.filterPlaceType !== undefined;
+
+    const coordsPlace = PlaceLocation.initFromLiteralCoords(searchTerm);
+    if (coordsPlace) {
+      this.resetMapPlaces();
+      this.handleSelectedPlace(coordsPlace);
+      return;
+    }
+    
+    this.fetchJourneyPoints(searchTerm);
   }
 
   onOpenAutocomplete() {
-    this.shouldFetchNewData = true
+    this.ignoreInputChanges = false;
   }
 
   onOptionSelected(ev: SbbAutocompleteSelectedEvent) {
-    this.shouldFetchNewData = false;
-
     const optionIdParts = ev.option.value.split('.');
     if (optionIdParts.length !== 2) {
       return;
@@ -109,19 +170,30 @@ export class JourneyPointInputComponent implements OnInit, OnChanges {
     
     const placeType = optionIdParts[0] as OJP_Types.PlaceTypeEnum;
     const itemIdx = parseInt(optionIdParts[1], 10);
+    const placeResult = this.mapLookupPlaces[placeType][itemIdx];
 
-    const place = this.mapLookupPlaces[placeType][itemIdx];
-    this.handleSelectedPlace(place);
+    this._currentRenderPlaceResult = placeResult;
+
+    if (placeResult.type === 'around_me') {
+      this.handleGeolocationLookup();
+    } else {
+      const place = placeResult.place;
+      this.handleSelectedPlace(place);
+    }
   }
 
   private async fetchJourneyPoints(searchTerm: string) {
-    const ojpSDK_Next = this.userTripService.createOJP_SDK_Instance(this.languageService.language);
-    const request = ojpSDK_Next.requests.LocationInformationRequest.initWithLocationName(searchTerm);
-    
-    const response = await request.fetchResponse(ojpSDK_Next);
+    const placeTypes = this.filterPlaceType === undefined ? undefined : [this.filterPlaceType];
+    const request = this.sdk.requests.LocationInformationRequest.initWithLocationName(searchTerm, placeTypes, 10);
+
+    await this.fetchRequest(request);
+  }
+
+  private async fetchRequest(request: AnyLocationInformationRequest) {
+    const response = await request.fetchResponse(this.sdk);
 
     if (!response.ok) {
-      console.log('ERROR - failed to lookup locations for "' + searchTerm + '"');
+      console.log('ERROR - failed to lookup locations');
       console.log(response);
       return;
     }
@@ -135,21 +207,27 @@ export class JourneyPointInputComponent implements OnInit, OnChanges {
           return;
       }
 
-      this.mapLookupPlaces[place.type].push(place);
+      const renderPlaceResult: RenderPlaceResult = {
+        place: place,
+        type: 'place',
+        caption: place.computeName(),
+      };
+
+      this.mapLookupPlaces[place.type].push(renderPlaceResult);
     });
   }
 
   public handleTapOnMapButton() {
-    if (this.currentPlace) {
-      this.mapService.tryToCenterAndZoomToPlace(this.currentPlace);
+    if (this.place) {
+      this.mapService.tryToCenterAndZoomToPlace(this.place);
     }
   }
 
   private handleSelectedPlace(place: AnyPlace) {
-    const inputValue = place.computeName();
-    this.inputControl.setValue(inputValue);
-
-    this.currentPlace = place;
+    // use ignoreInputChanges, otherwise emitEvent: false will not work with debounceTime()
+    this.ignoreInputChanges = true;
+    this.inputControl.setValue(place.computeName(), { emitEvent: false });
+    
     this.selectedNewPlace.emit(place);
   }
 
@@ -161,5 +239,83 @@ export class JourneyPointInputComponent implements OnInit, OnChanges {
       topographicPlace: [],
       location: [],
     };
+  }
+
+  private handleGeolocationLookup() {
+    if (!navigator.geolocation) {
+      console.error('no navigator.geolocation enabled')
+      return;
+    }
+    
+    // use ignoreInputChanges, otherwise emitEvent: false will not work with debounceTime()
+    this.ignoreInputChanges = true;
+    this.inputControl.setValue('... looking up location', { emitEvent: false });
+
+    navigator.geolocation.getCurrentPosition(
+      async position => {
+        await this.handleNewGeoPosition(position);
+
+        this.ignoreInputChanges = false;
+        this.inputControl.setValue('... choose nearby stop', { emitEvent: false });
+
+        this.autocompleteInputTrigger?.openPanel();
+      },
+      error => {
+        this.ignoreInputChanges = false;
+        this.inputControl.setValue('... Geolocation ERROR: ' + error.message, { emitEvent: false });
+
+        console.error('GeoLocation ERROR');
+        console.log(error);
+      },
+      {
+        enableHighAccuracy: false,
+        timeout: 15 * 1000,
+        maximumAge: 60 * 60 * 1000,
+      }
+    );
+  }
+
+  private async handleNewGeoPosition(position: GeolocationPosition) {
+    const nearbyGeoPosition = new OJP_Next.GeoPosition(position.coords.longitude, position.coords.latitude);
+    const bbox = GeoPositionBBOX.initFromGeoPosition(nearbyGeoPosition, 5000, 5000);
+    const bboxData = bbox.asFeatureBBOX();
+
+    const request = this.sdk.requests.LocationInformationRequest.initWithBBOX(bboxData, ['stop'], 300);
+
+    await this.fetchRequest(request);
+
+    const stopPlaces = this.mapLookupPlaces['stop'];
+    if (stopPlaces.length === 0) {
+      return;
+    }
+
+    stopPlaces.forEach(placeResult => {
+      placeResult.distance = placeResult.place.geoPosition.distanceFrom(nearbyGeoPosition);
+      placeResult.caption = placeResult.caption + ' (' + placeResult.distance + ' m)';
+    });
+
+    // Sort by distance if needed
+    stopPlaces.sort((a, b) => {
+      if (a.distance === null || a.distance === undefined) {
+        return 0;
+      }
+      if (b.distance === null || b.distance === undefined) {
+        return 0;
+      }
+      return a.distance - b.distance;
+    });
+
+    this.mapLookupPlaces['stop'] = stopPlaces.slice(0, 10);
+  }
+
+  public clearInputText() {
+    // use ignoreInputChanges, otherwise emitEvent: false will not work with debounceTime()
+    this.ignoreInputChanges = true;
+    this.inputControl.setValue('', { emitEvent: false });
+
+    this.resetMapPlaces();
+    this.mapLookupPlaces['stop'] = [AroundMePlaceResult];    
+    
+    this.autocompleteInputTrigger?.openPanel();
   }
 }