+++ src/app/shared/models/place/place-builder.ts
@@ -8,8 +8,9 @@ import { StopPlace } from './stop-place';
 import { Poi } from './poi';
 import { PlaceLocation } from './location';
 import { AnyPlaceResultSchema } from '../../types/_all';
+import { TopographicPlace } from './topographic-place';
 
-export type AnyPlace = Address | PlaceLocation | Poi | StopPlace;
+export type AnyPlace = Address | PlaceLocation | Poi | StopPlace | TopographicPlace;
 
 export class PlaceBuilder {
   public static initWithPlaceResultSchema(version: OJP_Next.OJP_VERSION, placeResultSchema: AnyPlaceResultSchema): AnyPlace | null {
@@ -44,6 +45,17 @@ export class PlaceBuilder {
     }
     if (placePoi) {
       return placePoi;
+    }
+
+    let placeTopographicPlace: TopographicPlace | null = null;
+    try {
+      placeTopographicPlace = TopographicPlace.initWithPlaceResultSchema(version, placeResultSchema);
+    } catch (error) {
+      console.error('TopographicPlace.initWithPlaceResultSchema:', error);
+      console.log(placeResultSchema);
+    }
+    if (placeTopographicPlace) {
+      return placeTopographicPlace;
     }
 
     return null;